Refactor dialect commons
Add Quote method to solve names escaping in the future
