+++ dialect_sqlite.go
@@ -33,9 +33,9 @@ func init() {
 var _ dialect = &sqlite{}
 
 type sqlite struct {
-	gil               *sync.Mutex
-	smGil             *sync.Mutex
-	ConnectionDetails *ConnectionDetails
+	commonDialect
+	gil   *sync.Mutex
+	smGil *sync.Mutex
 }
 
 func (m *sqlite) Name() string {
@@ -210,16 +210,12 @@ func (m *sqlite) TruncateAll(tx *Connection) error {
 	return tx.RawQuery(strings.Join(stmts, "; ")).Exec()
 }
 
-func (m *sqlite) afterOpen(c *Connection) error {
-	return nil
-}
-
 func newSQLite(deets *ConnectionDetails) (dialect, error) {
 	deets.URL = fmt.Sprintf("sqlite3://%s", deets.Database)
 	cd := &sqlite{
-		gil:               &sync.Mutex{},
-		smGil:             &sync.Mutex{},
-		ConnectionDetails: deets,
+		gil:           &sync.Mutex{},
+		smGil:         &sync.Mutex{},
+		commonDialect: commonDialect{ConnectionDetails: deets},
 	}
 
 	return cd, nil
